examples/aditof-demo: Fix memory leak by learing queue with frames
What was happening was that after clicking the stop button and then
click start again, the memory usage (on Windows) was increasing by 1.4
GB. And this could go on for each stop->start iteration.
The fix is to clear the queue with captured frames when stopping the
stream capture. Otherwise the next captuing will start with a large
queue (thinking it's empty) and appending new frames to that.

Signed-off-by: Dan Nechita <dan.nechita@analog.com>